pkg/blobserver/encrypt: Support the use of scrypt and a passphrase.
This commit adds the possiblity of using scrypt and a
password/passphrase to generate the key for the encrypt blobserver.

The old mechanism of using a hex-encoded 128 bit key in the
configuration is still supported and remains the default, but is
deprecated. If, in the encrypt blobserver configuration, "key" or
"keyFile" is defined, then that is what is used. Thus old
configurations and encrypt storages will continue to work as usual.

A notable feature of this change is that when the new mechanism of a
scrypt + a passphrase is used, a randomly generated 16 byte (128 bits)
salt is stored in and retrieved from the underlying blobstore. To this
end, we have implemented a new interface type, blobserver.Salter, and
only blobstores implementing this new interface can be used with the
scrypt/passphrase mechanism. At the moment, localdisk and diskpacked
implement the blobserver.Salter interface and they have been tested
with it.

Storing the the random salt in the blobstore allows a user,
in principle, to reconnect (from scratch) to a blobserver with
knowledge only of the password or passphrase.

If the new scrypt/passphrase mechanism is in use and the underlying
blobserver does not contain a salt, a random one is generated and
stored, but only if the blobserver is empty. This avoids the risk of
mixing blobs encrypted with different keys into the same underlying
blobserver. Switching an encrypt storage to the new mechanism
therefore requires specifying a new emtpy blobstore and syncing into
it.

Laslty, implementers of the blobserver.Salter interaface for composite
or wrapped storages such as replica or shard should take care that the
salts stored in the underlying storages are exactly the same.

Change-Id: I4f5224ba46e5cb2dcbf2eb0b48f4cfb83c04b54b